move server code under server subfolder
