+++ server/settings/publish_plugins.py
